requirements.txt
# FastAPI Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Authentication & Security
python-jose[cryptography]>=3.3.0
google-auth>=2.35.0
Authlib>=1.3.0
python-dotenv>=1.0.0

# Database & ORM
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# AI & Data Extraction
google-generativeai>=0.8.0
pypdf>=5.0.0
pymupdf>=1.24.0

# Data Validation
pydantic[email]>=2.9.0
pydantic-settings>=2.1.0

# PDF Generation & Assets
reportlab==4.0.9
pillow==10.2.0

# Utilities
python-dateutil==2.8.2
